TASK: Adjust migration for user interface mode
the migration previously adjusted the code to a now deprecated eel helper. In pr https://github.com/neos/neos-development-collection/pull/4505 the renderingMode is introduced as global fusionValue which made an adjustment in the rector necessary.

`(node|documentNode|site).context.inBackend` is now migrated to `renderingMode.isEdit`
`(node|documentNode|site).context.live` is now migrated to `!renderingMode.isEdit`

In cases where it cannot be determined wether a node is affected by the `context` operation a comment is added as before.